requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.3
pymysql
google-auth==2.25.2
google-auth-oauthlib==1.2.0
icalendar
python-dateutil
